+++ packages/eslint-plugin/src/rules/no-unnecessary-condition.ts
@@ -52,15 +52,9 @@ const isLiteral = (type: ts.Type): boolean =>
   isLiteralType(type);
 // #endregion
 
-type ExpressionWithTest =
-  | TSESTree.ConditionalExpression
-  | TSESTree.DoWhileStatement
-  | TSESTree.ForStatement
-  | TSESTree.IfStatement
-  | TSESTree.WhileStatement;
-
 export type Options = [
   {
+    allowConstantLoopConditions?: boolean;
     ignoreRhs?: boolean;
   },
 ];
@@ -87,6 +81,9 @@ export default createRule<Options, MessageId>({
       {
         type: 'object',
         properties: {
+          allowConstantLoopConditions: {
+            type: 'boolean',
+          },
           ignoreRhs: {
             type: 'boolean',
           },
@@ -108,10 +105,11 @@ export default createRule<Options, MessageId>({
   },
   defaultOptions: [
     {
+      allowConstantLoopConditions: false,
       ignoreRhs: false,
     },
   ],
-  create(context, [{ ignoreRhs }]) {
+  create(context, [{ allowConstantLoopConditions, ignoreRhs }]) {
     const service = getParserServices(context);
     const checker = service.program.getTypeChecker();
 
@@ -127,8 +125,16 @@ export default createRule<Options, MessageId>({
     function checkNode(node: TSESTree.Node): void {
       const type = getNodeType(node);
 
-      // Conditional is always necessary if it involves `any` or `unknown`
-      if (isTypeFlagSet(type, TypeFlags.Any | TypeFlags.Unknown)) {
+      // Conditional is always necessary if it involves:
+      //    `any` or `unknown` or a naked type parameter
+      if (
+        unionTypeParts(type).some(part =>
+          isTypeFlagSet(
+            part,
+            TypeFlags.Any | TypeFlags.Unknown | ts.TypeFlags.TypeParameter,
+          ),
+        )
+      ) {
         return;
       }
       const messageId = isTypeFlagSet(type, TypeFlags.Never)
@@ -197,14 +203,13 @@ export default createRule<Options, MessageId>({
      * Filters all LogicalExpressions to prevent some duplicate reports.
      */
     function checkIfTestExpressionIsNecessaryConditional(
-      node: ExpressionWithTest,
+      node: TSESTree.ConditionalExpression | TSESTree.IfStatement,
     ): void {
-      if (
-        node.test !== null &&
-        node.test.type !== AST_NODE_TYPES.LogicalExpression
-      ) {
-        checkNode(node.test);
+      if (node.test.type === AST_NODE_TYPES.LogicalExpression) {
+        return;
       }
+
+      checkNode(node.test);
     }
 
     /**
@@ -223,14 +228,46 @@ export default createRule<Options, MessageId>({
       }
     }
 
+    /**
+     * Checks that a testable expression of a loop is necessarily conditional, reports otherwise.
+     */
+    function checkIfLoopIsNecessaryConditional(
+      node:
+        | TSESTree.DoWhileStatement
+        | TSESTree.ForStatement
+        | TSESTree.WhileStatement,
+    ): void {
+      if (
+        node.test === null ||
+        node.test.type === AST_NODE_TYPES.LogicalExpression
+      ) {
+        return;
+      }
+
+      /**
+       * Allow:
+       *   while (true) {}
+       *   for (;true;) {}
+       *   do {} while (true)
+       */
+      if (
+        allowConstantLoopConditions &&
+        isBooleanLiteralType(getNodeType(node.test), true)
+      ) {
+        return;
+      }
+
+      checkNode(node.test);
+    }
+
     return {
       BinaryExpression: checkIfBinaryExpressionIsNecessaryConditional,
       ConditionalExpression: checkIfTestExpressionIsNecessaryConditional,
-      DoWhileStatement: checkIfTestExpressionIsNecessaryConditional,
-      ForStatement: checkIfTestExpressionIsNecessaryConditional,
+      DoWhileStatement: checkIfLoopIsNecessaryConditional,
+      ForStatement: checkIfLoopIsNecessaryConditional,
       IfStatement: checkIfTestExpressionIsNecessaryConditional,
-      WhileStatement: checkIfTestExpressionIsNecessaryConditional,
       LogicalExpression: checkLogicalExpressionForUnnecessaryConditionals,
+      WhileStatement: checkIfLoopIsNecessaryConditional,
     };
   },
 });