+++ packages/eslint-plugin/src/rules/no-untyped-public-signature.ts
@@ -106,7 +106,11 @@ export default util.createRule<Options, MessageIds>({
             });
           }
 
-          if (!isReturnTyped(node.value.returnType)) {
+          if (
+            node.kind !== 'constructor' &&
+            node.kind !== 'set' &&
+            !isReturnTyped(node.value.returnType)
+          ) {
             context.report({
               node,
               messageId: 'noReturnType',