+++ packages/eslint-plugin/tests/rules/func-call-spacing.test.ts
@@ -33,6 +33,39 @@ ruleTester.run('func-call-spacing', rule, {
       '( f )<a>( 0 )',
       '( (f) )<a>( (0) )',
       '( f()() )<a>(0)',
+
+      // optional call
+      'f?.();',
+      'f?.(a, b);',
+      'f?.b();',
+      'f?.b()?.c();',
+      'f.b?.();',
+      'f.b?.().c();',
+      'f()?.()',
+      '(function() {}?.())',
+      'f?.( (0) )',
+      '(function(){ if (foo) { bar(); } }?.());',
+      'f?.(0, (1))',
+      "describe/**/?.('foo', function () {});",
+      "describe?./**/('foo', function () {});",
+      '( f )?.( 0 )',
+      '( (f) )?.( (0) )',
+      '( f?.()() )(0)',
+      '( f()?.() )(0)',
+      '( f?.()?.() )(0)',
+      '( f?.()() )?.(0)',
+      '( f()?.() )?.(0)',
+      '( f?.()?.() )?.(0)',
+      'f?.<a>()',
+      'f?.<a>(b, b)',
+      'f.b?.<a>(b, b)',
+      'f?.b<a>(b, b)',
+      'f?.b?.<a>(b, b)',
+      '(function<T>() {}?.<a>())',
+      '((function<T>() {})<a>())',
+      '( f )?.<a>( 0 )',
+      '( (f) )?.<a>( (0) )',
+      '( f()() )?.<a>(0)',
     ].map<TSESLint.ValidTestCase<Options>>(code => ({
       code,
       options: ['never'],
@@ -60,6 +93,11 @@ ruleTester.run('func-call-spacing', rule, {
       '( f )<a> ( 0 )',
       '( (f) )<a> ( (0) )',
       '( f () )<a> (0)',
+
+      // optional call
+      'f?.b ();',
+      'f?.b ()?.c ();',
+      'f?.b<a> (b, b)',
     ].map<TSESLint.ValidTestCase<Options>>(code => ({
       code,
       options: ['always'],
@@ -76,6 +114,10 @@ ruleTester.run('func-call-spacing', rule, {
       'f\u2028();',
       'f\u2029();',
       'f\r\n();',
+
+      // optional call
+      'f?.b \n ();',
+      'f\n() ()?.b \n()\n ()',
     ].map<TSESLint.ValidTestCase<Options>>(code => ({
       code,
       options: ['always', { allowNewlines: true }],
@@ -364,5 +406,41 @@ var a = foo
           ...code,
         } as TSESLint.InvalidTestCase<MessageIds, Options>),
     ),
+
+    // optional chain
+    ...[
+      'f ?.();',
+      'f?. ();',
+      'f ?. ();',
+      'f\n?.();',
+      'f?.\n();',
+      'f\n?.\n();',
+    ].reduce<TSESLint.InvalidTestCase<MessageIds, Options>[]>((acc, code) => {
+      acc.push(
+        {
+          options: ['always', { allowNewlines: true }],
+          errors: [{ messageId: 'unexpected' }],
+          code,
+          // apply no fixers to it
+          output: null,
+        },
+        {
+          options: ['always'],
+          errors: [{ messageId: 'unexpected' }],
+          code,
+          // apply no fixers to it
+          output: null,
+        },
+        {
+          options: ['never'],
+          errors: [{ messageId: 'unexpected' }],
+          code,
+          // apply no fixers to it
+          output: null,
+        },
+      );
+
+      return acc;
+    }, []),
   ],
 });