+++ packages/eslint-plugin/tests/rules/no-floating-promises.test.ts
@@ -219,6 +219,18 @@ async function test() {
   return promise;
 }
 `,
+
+    // optional chaining
+    `
+async function test() {
+  declare const returnsPromise: () => Promise<void> | null;
+  await returnsPromise?.();
+  returnsPromise()?.then(() => {}, () => {});
+  returnsPromise()?.then(() => {})?.catch(() => {});
+  returnsPromise()?.catch(() => {});
+  return returnsPromise();
+}
+  `,
   ],
 
   invalid: [